refactor: move cancellation helpers out of constants
